Remove `sync-team` CLI and call it from `team` as a library
