+++ Cargo.toml
@@ -22,6 +22,8 @@ serde_json = "1"
 serde-untagged = "0.1"
 toml = "0.8"
 
+sync-team = { path = "sync-team" }
+
 [dev-dependencies]
 ansi_term = "0.12.1"
 atty = "0.2.14"
