+++ sync-team/Cargo.toml
@@ -5,10 +5,8 @@ authors = ["Pietro Albini <pietro@pietroalbini.org>"]
 edition = "2024"
 
 [dependencies]
-clap = { version = "4.5", features = ["derive", "env"] }
 reqwest = { version = "0.12.8", features = ["blocking", "json", "rustls-tls", "charset", "http2", "macos-system-configuration"], default-features = false }
 log = "0.4"
-env_logger = "0.11"
 rust_team_data = { path = "../rust_team_data", features = ["email-encryption"] }
 serde = { version = "1.0", features = ["derive"] }
 anyhow = "1.0"
